AUTHLIB-161 Remove Deprecated Code
Remove code deprecated in version 2.3.0:

* Remove email configuration replaced by messaging-lib properties
* Remove view helper classes replaced by common-lib equivalents
* Refactor to use `AbstractEntity` from common-lib
