+++ authentication_lib/src/main/java/org/octri/authentication/server/security/entity/PasswordResetToken.java
@@ -3,12 +3,11 @@ package org.octri.authentication.server.security.entity;
 import java.util.Date;
 import java.util.UUID;
 
+import org.octri.common.domain.AbstractEntity;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Transient;
@@ -21,57 +20,45 @@ import jakarta.validation.constraints.NotNull;
  * @author sams
  */
 @Entity
-public class PasswordResetToken {
+public class PasswordResetToken extends AbstractEntity {
 
-	@Id
-	@Column(name = "id")
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
+	/**
+	 * Unique token string.
+	 */
 	@NotNull
 	@Column(name = "token", unique = true)
 	private String token = UUID.randomUUID().toString();
 
+	/**
+	 * The user whose password will be reset.
+	 */
 	@NotNull
 	@OneToOne(targetEntity = User.class, fetch = FetchType.EAGER)
 	@JoinColumn(nullable = false, name = "user")
 	private User user;
 
+	/**
+	 * Date and time when the token will expire.
+	 */
 	@NotNull
 	@Column(name = "expiry_date")
 	private Date expiryDate;
 
+	/**
+	 * Transient property used to store the password reset URL.
+	 */
 	@Transient
 	private String tokenUrl;
 
 	/**
-	 * Default contructor, no fields are set.
+	 * Default constructor, no fields are set.
 	 */
 	public PasswordResetToken() {
 	}
 
 	/**
-	 * Gets the record's unique ID.
-	 * 
-	 * @return the unique ID
-	 */
-	public Long getId() {
-		return id;
-	}
-
-	/**
-	 * Sets the record's unique ID.
-	 * 
-	 * @param id
-	 *            the ID to set
-	 */
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	/**
 	 * Gets the token string.
-	 * 
+	 *
 	 * @return the token string
 	 */
 	public String getToken() {
@@ -80,7 +67,7 @@ public class PasswordResetToken {
 
 	/**
 	 * Sets the token string.
-	 * 
+	 *
 	 * @param token
 	 *            the token string to set
 	 */
@@ -90,7 +77,7 @@ public class PasswordResetToken {
 
 	/**
 	 * Gets the user account that can be reset using the token.
-	 * 
+	 *
 	 * @return the associated user account
 	 */
 	public User getUser() {
@@ -99,7 +86,7 @@ public class PasswordResetToken {
 
 	/**
 	 * Sets the user account that can be reset using the token.
-	 * 
+	 *
 	 * @param user
 	 *            a user account
 	 */
@@ -109,7 +96,7 @@ public class PasswordResetToken {
 
 	/**
 	 * Gets the timestamp when the token expires and can no longer be used.
-	 * 
+	 *
 	 * @return when the token expires
 	 */
 	public Date getExpiryDate() {
@@ -118,7 +105,7 @@ public class PasswordResetToken {
 
 	/**
 	 * Sets the timestamp when the token expires and can no longer be used.
-	 * 
+	 *
 	 * @param expiryDate
 	 *            when the token should expire
 	 */
@@ -128,7 +115,7 @@ public class PasswordResetToken {
 
 	/**
 	 * Gets the URL that initiates a password reset using the token.
-	 * 
+	 *
 	 * @return password reset URL
 	 */
 	public String getTokenUrl() {
@@ -137,7 +124,7 @@ public class PasswordResetToken {
 
 	/**
 	 * Sets the URL that initiates a password reset using the token.
-	 * 
+	 *
 	 * @param tokenUrl
 	 *            password reset URL
 	 */
