--- authentication_lib/src/main/java/org/octri/authentication/server/view/OptionList.java
@@ -1,90 +1,0 @@
-package org.octri.authentication.server.view;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.StreamSupport;
-
-/**
- * Used for rendering mustache templates. Helper functions for creating a list of select input options.
- *
- * @deprecated
- *             Use {@link org.octri.common.view.OptionList} from
- *             <a href="https://github.com/OHSU-OCTRI/common-lib">OCTRI common-lib</a> class instead.
- *
- * @author lawhead
- * @param <T>
- *            the type of objects in the select list
- *
- */
-@Deprecated(since = "2.3.0", forRemoval = true)
-public class OptionList<T> {
-
-	/**
-	 * Given a Repository search result of lookups and the selected lookup item, provides a list of objects that can be
-	 * used directly by mustachejs for rendering.
-	 *
-	 * @deprecated
-	 *             Use
-	 *             {@link org.octri.common.view.OptionList#fromSearch(Iterable, org.octri.common.domain.AbstractEntity)}
-	 *             instead.
-	 *
-	 * @param <T>
-	 *            a type with an ID and label
-	 * @param iter
-	 *            an iterable collection
-	 * @param selected
-	 *            the current selection (may be null)
-	 * @return a list of select options for the items in the collection
-	 */
-	public static <T extends Identified & Labelled> List<EntitySelectOption<T>> fromSearch(Iterable<T> iter,
-			T selected) {
-		return StreamSupport.stream(iter.spliterator(), false)
-				.map(area -> new EntitySelectOption<T>(area, selected))
-				.collect(Collectors.toList());
-	}
-
-	/**
-	 * Used for multi-selects. Given a Repository search result of lookups and a list of selected lookup, provides a
-	 * list of objects that can be used directly by mustachejs for rendering.
-	 *
-	 * @deprecated
-	 *             Use {@link org.octri.common.view.OptionList#multiFromSearch(Iterable, Collection)} instead.
-	 *
-	 * @param <T>
-	 *            a type with an ID and a label
-	 * @param iter
-	 *            an iterable collection
-	 * @param selected
-	 *            the current selections (may be empty)
-	 * @return a list of select options for the items in the collection
-	 */
-	public static <T extends Identified & Labelled> List<EntitySelectOption<T>> multiFromSearch(Iterable<T> iter,
-			Collection<T> selected) {
-		return StreamSupport.stream(iter.spliterator(), false)
-				.map(area -> new EntitySelectOption<T>(area, selected))
-				.collect(Collectors.toList());
-	}
-
-	/**
-	 * Generates a list of integers in the given range from which to choose.
-	 *
-	 * @deprecated
-	 *             Use {@link org.octri.common.view.OptionList#forRange(Integer, Integer, Integer)} instead.
-	 *
-	 * @param start
-	 *            start of the range
-	 * @param end
-	 *            end of the range
-	 * @param selected
-	 *            the current selection (may be null)
-	 * @return a list of select options for the integers in the range
-	 */
-	public static List<SelectOption<Integer>> forRange(Integer start, Integer end, Integer selected) {
-		return IntStream.rangeClosed(start, end)
-				.mapToObj(i -> new SelectOption<Integer>(Integer.valueOf(i), selected))
-				.collect(Collectors.toList());
-	}
-
-}
