+++ authentication_ui_bootstrap5/src/main/resources/mustache-templates/authlib_fragments/admin/user/form.mustache
@@ -100,6 +100,7 @@
 							<input class="form-control" value="{{user.username}}" type="text" disabled>
 							<input type="hidden" id="username" name="username" value="{{user.username}}">
 							<input class="form-control" type="hidden" id="id" name="id" value="{{user.id}}">
+							<input type="hidden" id="version" name="version" value="{{user.version}}">
 						</div>
 
 						<div class="form-class">
